More files added, compiling fine now
So, there were a few things that were fixed. One, Windows automatic file
extensions were messing things up. The main components are just about
done, (that is the root  nodes of each inheritance tree, small as they
are). We have a game class, we have a scene, a render engine, a
game_object, and on derived game object (the ship. For testing). I've
implemented crude signalling between the game_objects and the scene in
order to have game_objects spawning other objects (ie an enemy spawner
object might use this or a ship might spawn bullets).

Next, we need to consider a few things. Physics engine needs to be
implemented to, which should be able to iterate over game_objects ina
scene and check for trigger/collider collisions and entry. This need not
be a true signal system, but the method on the game object might look a
bit like on_trigger_enter(Collider other_collider). Each collider that
interacts with both mutually have such a method called on them (as they
both sort of "enter" each other's collider).

Next as well, input management needs to be taken care of, and will work
similiar to the physics engine, except whereas the physics engine houses
all of the physics logic, the input engine will simply pass raw inputs
to the game_object to imlement it's own logic.

Finally, we are still lacking for implementations of classes such as
Text, the main menu, and the leaderboard, as well as enemy
implementation, enemy spwaner implementation, and a logic in the level
class for things such as current score and such.

In conclusion, much of the course, high level stuff *should* be about
finished by now, and it's detail that we need to implement now.  That
said, no teseting other than for compilation has occurred yet, because
hardware fabrication is still in the works. It's getting there though,
should be done in the next day or so.