Adjusted parameters to fit new board and refactored
I've gone ahead and updated the paramters in the render engine to
properly connect to our new board.

Did some light testing as well, but the screen may be a dud, which is
not good considering the expo is tomorrow. As well, I knew I would
regret our "architecture" sooner rather than later. Note to self, in the
future, whenever you here anything along the lines of "it's a small
project, let's just cheat and make the 'easy' architecture. It doesn't
need to be maintainable in this case.", you better punch whoever said
that.

This code is only a few hundred lines at most, and everything is barely
hanging together. I imagine breathing too hard would cause the build to
break. Circular dependencies all over the place, dodged by using ugly
forward declarations and including multiple headers in implementation
files is the result.  This needs to be fixed very badly, but I have
neither the time before the expo, or the willingness to right now.