Input management is in place now
Implemented interrupt based input management. Note that the input
managmenet is handled in a persistent fashion, by which I mean that the
class is virtually entirely static. this is because arduino interrupts
must be static methods, which implies that the state of the input
manager must also be static, so there's no point in having an instanced
manager.