Quick backup of newly implemented systems.
A lot has been added, so now is as good a time as any to backup. We
added objects to the game including bullets and enemies.  We've also
finalized our signaling system, and implemented sound collission system
(well mostly sound, we don't check for collisions yet, but collision
signalling does indeed exist right now). The screen also works, which is
a huge leg up, and bugs have been removed left and right.

At this point, there's only a handful of things left to  implemeent
before the game is playable. These include actually checking for
collissions, UI (i.e. for score and number of lives left), and we need
to hook up the control circuit for the controller input (i.e. move &
shoot buttons) for proper testing. Then we'll have our game finished. if
we have time later, perhaps a menu would be nice, and much farther down
if we feel like it, a leaderboard.

Note the leaderboard is not pressing since we don't have an SD card
arduino shield, so scores right now would only be held in RAM until the
arduino were restarted anwyay .