Finished the level and functional level-selection
The game is now playable and the game can select levels to load. The
player starts with three lives and plays until the end, where the game
then displays game over and the arduino must be reset manually.

Obviously, a software reset would be nice, as well as perhaps a welcome
screen. Though that's icing on the cake considering the deadline is in
less than 24 hours.

As well, there are some kinks that should be worked out sooner rather
than later. One, the screen flicker is absurd. I don't know what causes
it, but i'll just say, if you're epileptic, you're gonna have a bad time
playing this. As well, the controls for the ship are largely untested. A
few light tests were conudcted earlier, but more thorough testing and
implementation should be done.

As well, enemy ships die on friendly fire, which isn't ideal. A simple
bullet direction check could fix that though.

Finally, we need to implement invulnerability for hte ship on respawn.
The ship dies, then respawns, and if the ship died at the same spot it
respawns immediately, then the ship get's killed again by the same
bullet that killed it's first life.

Also, an easy method for passing information between levels would be
nice (i.e. displaying the score earned on teh game over screen.). Of
course, this requires a more involved refactoring of game->level
signalling that would require more work than it's worth at this point in
time, but I'll put it there because it might be nice to add as a feature
later.

Finally, data compression needs to happen in a lot of spots. There's a
lot of unnecessarily large data types (i.e. int where an int8_t would
suffice). We're up to 50% RAM usage, so this is a non-useless fix. As
well, some of our algorithms stand to be enhanced. Perhaps sorting
game_objects according to slots and such would allow for quicker updates
as we can use quick_search rather than linear search for deleting,
updating, etc. As well, it would enhance speed based on
branch-predictors. Then of course there's the matter of physics, which
is currently O(n^2), which makes me cry on the inside every time I look
at it. Perhaps a rough distance-based cull before brute-force checking
collision would lessen the number of comparisons.